mv: terminal function to separate file
